# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepriseAI Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/agent/login": {
      "post": {
        "tags": ["Agent"],
        "summary": "Agent Login",
        "description": "Agent login endpoint.",
        "operationId": "agent_login_agent_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentToken"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/me": {
      "get": {
        "tags": ["Agent"],
        "summary": "Get Current Agent Profile",
        "description": "Get current agent's profile with partner hold status.",
        "operationId": "get_current_agent_profile_agent_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentNameOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/agent/orders": {
      "get": {
        "tags": ["Agent"],
        "summary": "Get Agent Orders",
        "description": "Get all orders assigned to the current agent.",
        "operationId": "get_agent_orders_agent_orders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by order status",
              "title": "Status Filter"
            },
            "description": "Filter by order status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Agent Orders Agent Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/orders/{order_id}": {
      "get": {
        "tags": ["Agent"],
        "summary": "Get Agent Order Detail",
        "description": "Get detailed information about a specific order.",
        "operationId": "get_agent_order_detail_agent_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Agent Order Detail Agent Orders  Order Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/orders/{order_id}/complete-pickup": {
      "post": {
        "tags": ["Agent"],
        "summary": "Complete Pickup",
        "description": "Complete the pickup process with detailed phone inspection.\nRecords actual condition, final price, customer acceptance, and photos.\nStores all photos as binary BLOB data in the database.",
        "operationId": "complete_pickup_agent_orders__order_id__complete_pickup_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_complete_pickup_agent_orders__order_id__complete_pickup_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/orders/{order_id}/process-payment": {
      "post": {
        "tags": ["Agent"],
        "summary": "Process Payment",
        "description": "Process payment for a completed pickup.\nOnly available if customer accepted the offer.",
        "operationId": "process_payment_agent_orders__order_id__process_payment_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/orders/{order_id}/reschedule-pickup": {
      "post": {
        "tags": ["Agent"],
        "summary": "Reschedule Pickup",
        "description": "Reschedule a pickup that was previously scheduled.",
        "operationId": "reschedule_pickup_agent_orders__order_id__reschedule_pickup_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReschedulePickupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/orders/{order_id}/cancel-pickup": {
      "post": {
        "tags": ["Agent"],
        "summary": "Cancel Pickup",
        "description": "Cancel a pickup when customer is not willing to sell.\nMarks the order as cancelled (terminal state).",
        "operationId": "cancel_pickup_agent_orders__order_id__cancel_pickup_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelPickupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": ["email", "password"],
        "title": "AgentLogin",
        "description": "Schema for agent login"
      },
      "AgentNameOut": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "is_on_hold": {
            "type": "boolean",
            "title": "Is On Hold",
            "default": false
          },
          "hold_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Reason"
          },
          "hold_lift_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Lift Date"
          }
        },
        "type": "object",
        "required": ["full_name"],
        "title": "AgentNameOut",
        "description": "Schema for agent name and hold status"
      },
...
      "AgentToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "agent": {
            "$ref": "#/components/schemas/backend__services__partner__schema__schemas__AgentOut"
          }
        },
        "type": "object",
        "required": ["access_token", "agent"],
        "title": "AgentToken",
        "description": "Schema for agent authentication token"
      },
...
      "Body_complete_pickup_agent_orders__order_id__complete_pickup_post": {
        "properties": {
          "final_offered_price": {
            "type": "number",
            "title": "Final Offered Price"
          },
          "customer_accepted": {
            "type": "boolean",
            "title": "Customer Accepted"
          },
          "actual_condition": {
            "type": "string",
            "title": "Actual Condition",
            "default": "Inspected"
          },
          "pickup_notes": {
            "type": "string",
            "title": "Pickup Notes",
            "default": ""
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "default": ""
          },
          "phone_conditions": {
            "type": "string",
            "title": "Phone Conditions"
          },
          "photos": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": ["final_offered_price", "customer_accepted"],
        "title": "Body_complete_pickup_agent_orders__order_id__complete_pickup_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/admin/auth/login"
          }
        }
      }
    }
  }
}